Always evaluate SqlExpresion before accessing params
