+++ tests/ServiceStack.OrmLite.Tests/Issues/LoadSelectIssue.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Data;
 using NUnit.Framework;
 using ServiceStack.Data;
 using ServiceStack.DataAnnotations;
@@ -130,6 +131,15 @@ namespace ServiceStack.OrmLite.Tests.Issues
             bool IsDeleted { get; set; }
         }
 
+        private static void CreateTables(IDbConnection db)
+        {
+            db.DropTable<EventCategoryTbl>();
+            db.DropTable<FileTbl>();
+
+            db.CreateTable<FileTbl>();
+            db.CreateTable<EventCategoryTbl>();
+        }
+
         [Test]
         public void Can_execute_LoadSelect_when_child_references_implement_IHasSoftDelete()
         {
@@ -145,13 +155,38 @@ namespace ServiceStack.OrmLite.Tests.Issues
 
             using (var db = OpenDbConnection())
             {
-                db.DropTable<EventCategoryTbl>();
-                db.DropTable<FileTbl>();
-
-                db.CreateTable<FileTbl>();
-                db.CreateTable<EventCategoryTbl>();
+                CreateTables(db);
 
                 var results = db.LoadSelect<EventCategoryTbl>();
+            }
+
+            OrmLiteConfig.SqlExpressionSelectFilter = null;
+        }
+
+        [Test]
+        public void Can_execute_SoftDelete_with_GroupBy()
+        {
+            OrmLiteConfig.SqlExpressionSelectFilter = q =>
+            {
+                if (q.ModelDef.ModelType.HasInterface(typeof(IHasSoftDelete)))
+                {
+                    q.Where<IHasSoftDelete>(x => x.IsDeleted != true);
+                }
+            };
+
+            using (var db = OpenDbConnection())
+            {
+                CreateTables(db);
+
+                var name = "name";
+                var q = db.From<FileTbl>()
+                    .Where(x => x.Name == name && x.FileSizeBytes > 1000)
+                    .GroupBy(x => x.Extension)
+                    .Select(x => new { x.Extension, Total = Sql.As(Sql.Count("*"), "Total") });
+
+                var results = db.Dictionary<int, long>(q);
+
+                results.PrintDump();
             }
 
             OrmLiteConfig.SqlExpressionSelectFilter = null;
